requirements.txt
google-adk>=1.0.0
google-genai>=1.5.0
streamlit>=1.41.0
yfinance>=0.2.50
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0
pydantic>=2.10.0
requests>=2.32.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
lxml>=5.3.0
nest-asyncio>=1.6.0
